Rudimentary Collisions Set
Currently we need to change the projection since the output coordinates are normalized device coordinates (ndc). The width is in pixel units at the moment, but we need to use the vertex shader in order to transform positions into clip coordinates. Since there are simple gases, then there are only simple AABB collisions at the moment and perfectly elastic collisions.